dev-requirements.txt
coverage
pytest-django
pytest-randomly
selenium
